add email service tests, move mocks, change eslint
