+++ package.json
@@ -78,7 +78,6 @@
     "eslint-plugin-prettier": "^5.0.0",
     "eslint-plugin-security": "^1.7.1",
     "jest": "^29.7.0",
-    "nodemailer-mock": "^2.0.2",
     "prettier": "^3.0.0",
     "source-map-support": "^0.5.21",
     "supertest": "^6.3.3",
