+++ src/auth/test/auth.service.spec.ts
@@ -4,13 +4,13 @@ import { AuthService } from '../auth.service';
 import { UsersService } from '../../users/users.service';
 import { JwtService } from '@nestjs/jwt';
 import { ConfigService } from '@nestjs/config';
-import { usersServiceMock } from '../../utils/mocks/services/users.service.mock';
-import { jwtServiceMock } from '../../utils/mocks/services/jwt.service.mock';
-import { emailServiceMock } from '../../utils/mocks/services/email.service.mock';
-import { createUserDTOMock } from '../../utils/mocks/dtos/create-user.dto.mock';
-import { userSignInDTOMock } from '../../utils/mocks/dtos/user-sign-in.dto.mock';
+import { usersServiceMock } from './mocks/users.service.mock';
+import { jwtServiceMock } from './mocks/jwt.service.mock';
+import { emailServiceMock } from './mocks/email.service.mock';
+import { createUserDTOMock } from '../../utils/shared-mocks/create-user.dto.mock';
+import { userSignInDTOMock } from './mocks/user-sign-in.dto.mock';
 import { User } from '../../users/user.entity';
-import { tokenMock } from '../../utils/mocks/tokens/token.mock';
+import { tokenMock } from '../../utils/shared-mocks/token.mock';
 import { hash } from 'bcrypt';
 import { EmailService } from '../../email/email.service';
