--- src/utils/mocks/dtos/user.dto.mock.ts
@@ -1,11 +1,0 @@
-export const userDTOMock = {
-  login: 'test@test.test',
-  firstName: 'test',
-  lastName: 'test',
-  birthday: '2000-01-01',
-  country: 'test',
-  city: 'test',
-  postalCode: 'test',
-  street: 'test',
-  buildingNumber: 'test',
-};
